+++ ui-tests/jupyter_server_test_config.py
@@ -9,7 +9,7 @@ from jupyterlab.galata import configure_jupyter_server
 configure_jupyter_server(c)
 
 c.ServerApp.jpserver_extensions = {
-    "jupyterlab_unfold": True,
+    "jupyterlab_speedy_unfold": True,
 }
 
 # Uncomment to set server log level to debug level